requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
distlib==0.3.7
Django==5.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.0
gunicorn==21.2.0
packaging==24.0
pillow==10.2.0
platformdirs==3.11.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
virtualenv==20.24.6
whitenoise==6.6.0
